refactor(translator): decouple client deps + extract service entrypoints (A3–A5)
Step A3: Migrate Flow360TranslationError and Flow360ErrorWithLocation to
flow360_schema.exceptions (stdlib logging, not flow360.log). Client
flow360/exceptions.py re-exports them so existing callers keep working.

Delete client-side bet_translator_interface.py — schema has an
identical copy (functionally equivalent, only style differences).
services.py now imports directly from flow360_schema.

Step A4: In surface_meshing_translator.py, replace `from flow360.log
import log` with stdlib `logging.getLogger(__name__)`. Bridge in
flow360/log.py stays — it serves the flow360_schema.* namespace
(schema runs inside the client process). Translator runs only in
backend pipeline/webAPI/test contexts where rich logger integration
is unnecessary; callers configure their own root handlers.

Step A5: Extract the four `simulation_to_*_json` wrappers plus the
`_translate_simulation_json` helper from services.py into a new
translator/service_entrypoints.py. services.py re-exports them for
backward compat. Non-translator code (validate_model,
generate_process_json, unit-system conversion, BET webAPI entries)
stays in services.py.

Pre-existing test failures (4 in test_solver_translator.py, 5 in
test_translator_service.py, 8 errors in test_gai_coordinate_system_e2e.py)
are unchanged by these steps — verified via stash-rerun.
